% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calib_plot.R
\name{binary_calibration_plot}
\alias{binary_calibration_plot}
\title{Binary Calibration Plots}
\usage{
binary_calibration_plot(response, model, breaks = 0:10/10,
  point.color = "blue", line.color = "red")
}
\arguments{
\item{response}{The response variable from the data}

\item{model}{A model object}

\item{breaks}{The number of bins used to create the error bars that will be displayed}

\item{point.color}{The color of the points in the scatterplot}

\item{line.color}{The color of the line y = x}
}
\value{
A ggplot2 object
}
\description{
Binary Calibration Plots
}
\examples{
ch <- read.csv("http://www.stat.cmu.edu/~cshalizi/uADA/15/hw/06/ch.csv")
ch.clean <- na.omit(ch)
lr.ch <- glm(start ~ . -country -year +I(exports^2), data=ch, family="binomial") # logistic regression
binary_calibration_plot(y = ch.clean$start, model = lr.ch, breaks = 0:7/10, point.color = "blue", line.color = "red")
library(mgcv)
gam.ch <- gam(start~s(exports)+s(schooling)+s(growth)+s(peace)+s(concentration)+s(lnpop)+s(fractionalization)+dominance, # GAM
binary_calibration_plot(ch.clean$start, gam.ch, breaks=0:9/10, blue, red)
}
