% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostrap_ci.R
\name{bootstrap_ci}
\alias{bootstrap_ci}
\title{Bootstrap confidence interval}
\usage{
bootstrap_ci(statistic = NULL, simulator = NULL, tboots = NULL,
  B = if (!is.null(tboots)) {     ncol(tboots) }, t.hat, level)
}
\arguments{
\item{level}{The specified confidence level}
}
\value{
A confidence interval that was generated
}
\description{
Bootstrap confidence interval
}
\examples{


}
