% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loocv.R
\name{loocv}
\alias{loocv}
\title{Leave one out cross validation}
\usage{
loocv(model)
}
\arguments{
\item{model}{An R object, typically returned by lm or glm.}
}
\value{
Cross validation risk error
}
\description{
Leave one out cross validation
}
